feat: syncer

+++ src/node/state.rs
@@ -11,7 +11,7 @@ pub trait State {
     fn latest_block_hash(&self) -> Hash;
     fn latest_block_number(&self) -> u64;
     fn add_block(&mut self, block: Block) -> Result<Hash, ChainError>;
-    fn get_blocks(&self, offset: usize) -> Result<Vec<Block>, ChainError>;
+    fn get_blocks(&self, offset: u64) -> Result<Vec<Block>, ChainError>;
     fn get_block(&self, number: u64) -> Result<Block, ChainError>;
     fn get_mining_difficulty(&self) -> usize;
 }
